added animation and ui fixes

+++ xuapi/src/main/java/com/android/xu/xuapi/entities/Current.kt
@@ -24,7 +24,7 @@ data class Current (
 	@SerializedName("condition") val condition : Condition,
 	@SerializedName("wind_mph") val wind_mph : Double,
 	@SerializedName("wind_kph") val wind_kph : Double,
-	@SerializedName("wind_degree") val wind_degree : Int,
+	@SerializedName("wind_degree") val wind_degree : Double,
 	@SerializedName("wind_dir") val wind_dir : String,
 	@SerializedName("pressure_mb") val pressure_mb : Int,
 	@SerializedName("pressure_in") val pressure_in : Double,
@@ -34,9 +34,9 @@ data class Current (
 	@SerializedName("cloud") val cloud : Int,
 	@SerializedName("feelslike_c") val feelslike_c : Double,
 	@SerializedName("feelslike_f") val feelslike_f : Double,
-	@SerializedName("vis_km") val vis_km : Int,
-	@SerializedName("vis_miles") val vis_miles : Int,
-	@SerializedName("uv") val uv : Int,
+	@SerializedName("vis_km") val vis_km : Double,
+	@SerializedName("vis_miles") val vis_miles : Double,
+	@SerializedName("uv") val uv : Double,
 	@SerializedName("gust_mph") val gust_mph : Double,
 	@SerializedName("gust_kph") val gust_kph : Double
 )
